Modularize the multiprocess server code
